Fix warnings

+++ Partisans/Classes/SetupGameMenuItems.m
@@ -416,7 +416,7 @@
     return nil;
 }
 
-- (float)menuViewController:(JSKMenuViewController *)menuViewController heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (double)menuViewController:(JSKMenuViewController *)menuViewController heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     if (indexPath.section == SetupGameMenuSectionGame && indexPath.row == SetupGameMenuRowPlayers)
     {
